ostep project: wish, outline of main loop
Here is an outline of the main program loop. The focus at this stage was
very much on interactive mode. It contains an implementation of a very
simple parser, which tokenizes the any input coming in via stdin. A
token is simply any word suffixed with a space delimeter.

In interactive mode, the shell can handle multi-line reads,
i.e. once it reads a '\n', it simply replaces this with '\0' and keeps
reading. It pushes each token onto a argument stack that is being resized
according to the number of tokens being read.